requirements.txt
fastapi==0.85.0
uvicorn==0.17.6
passlib==1.7.4
bcrypt==3.2.0
pydantic==1.9.0
python-jose==3.3.0
pymongo==4.0.1
email-validator==1.1.3
python-dotenv
